make maven coordinates immutable again
