+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries.test/src/com/google/cloud/tools/eclipse/appengine/libraries/model/MavenCoordinatesTest.java
@@ -17,8 +17,6 @@
 package com.google.cloud.tools.eclipse.appengine.libraries.model;
 
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
@@ -164,16 +162,5 @@ public class MavenCoordinatesTest {
         .build();
     assertTrue(mavenCoordinates.getClassifier().isEmpty());
   }
-  
-  @Test
-  public void testUpdateVersion() {
-    MavenCoordinates mavenCoordinates = new MavenCoordinates.Builder()
-        .setGroupId("com.google.guava")
-        .setArtifactId("guava")
-        .setVersion("15.0")
-        .build();
-    mavenCoordinates.updateVersion();
-    assertFalse(mavenCoordinates.getVersion().isEmpty());
-    assertNotEquals("15.0", mavenCoordinates.getVersion());
-  }
+
 }
