+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/src/com/google/cloud/tools/eclipse/appengine/libraries/model/MavenCoordinates.java
@@ -18,9 +18,6 @@ package com.google.cloud.tools.eclipse.appengine.libraries.model;
 
 import java.text.MessageFormat;
 
-import org.apache.maven.artifact.versioning.ArtifactVersion;
-
-import com.google.cloud.tools.eclipse.util.ArtifactRetriever;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 
@@ -39,16 +36,13 @@ public class MavenCoordinates {
   private String version = LATEST_VERSION;
   private String type = JAR_TYPE;
   private String classifier;
-  
-  /** Version has been checked against Maven Central **/
-  // todo could mark this transient and on deserialization set depending on whether
-  // version is still LATEST_VERSION or not
-  private boolean fixedVersion = false;
 
   /**
    * @param groupId the Maven group ID, cannot be <code>null</code>
    * @param artifactId the Maven artifact ID, cannot be <code>null</code>
    */
+  // that this is public is only for legacy tests. We can easily convert those to use
+  // the builder instead
   @VisibleForTesting
   public MavenCoordinates(String groupId, String artifactId) {
     Preconditions.checkNotNull(groupId, "groupId null");
@@ -185,8 +179,8 @@ public class MavenCoordinates {
     }
 
     /**
-     * @param version the Maven version of the artifact, defaults to special value
-     *     {@link MavenCoordinates#LATEST_VERSION}, cannot be <code>null</code> or empty string.
+     * @param version the Maven version of the artifact. Defaults to special value
+     *     {@link MavenCoordinates#LATEST_VERSION}. Cannot be <code>null</code> or empty string.
      */
     public Builder setVersion(String version) {
       Preconditions.checkNotNull(version, "version is null");
@@ -204,21 +198,5 @@ public class MavenCoordinates {
     }
     
   }
-
-  /**
-   * Check Maven Central to find the latest release version of this artifact.
-   * This check is made at most once. Subsequent checks are no-ops.
-   */
-  public void updateVersion() {
-    if (!fixedVersion) {
-      // todo need method to get latest nonrelease version instead for alphas and betas
-      ArtifactVersion remoteVersion =
-          ArtifactRetriever.DEFAULT.getLatestArtifactVersion(groupId, artifactId);
-      if (remoteVersion != null) {
-        this.version = remoteVersion.toString();  
-      }
-      fixedVersion = true;
-    }
-  }  
 
 }
